+++ src/main/java/oncall/dto/ScheduleOutputDto.java
@@ -1,6 +1,6 @@
 package oncall.dto;
 
-import oncall.model.Week;
+import oncall.model.date.Week;
 
 public record ScheduleOutputDto(Integer month, Integer day, Week week, boolean isHoliday, String workerName) {
 }
